+++ NEWS.md
@@ -14,6 +14,8 @@
   (#516).
 - LLVM 8.0 or later is now required due to deprecation of non-opaque
   pointers.
+- Altera/Intel Quartus vendor libraries can now be compiled with `nvc
+  --install quartus`.
 
 ## Version 1.7.0 - 2022-08-07
 - *Breaking change:* In-tree builds are no longer supported: use a